requirements.txt
click==8.1.3
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
PyYAML==6.0.1
jsonschema==4.22.0
Deprecated==1.2.14
importlib_metadata==8.4.0
wrapt==1.16.0
zipp==3.20.2
grpcio==1.66.1
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opentelemetry-exporter-zipkin-json==1.27.0
opentelemetry-exporter-zipkin-proto-http==1.27.0
opentelemetry-exporter-prometheus==0.48b0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
